+++ extensions/readwise-to-tana/src/index.tsx
@@ -37,17 +37,20 @@ export default function Command() {
 
   const handleSave = async (values: SettingsValues) => {
     const {
-      supertag,
       author,
       category,
       coverImageUrl,
+      highlightColor,
+      highlightHighlightedAt,
+      highlightLocation,
       highlightNote,
       highlightSupertag,
-      highlightLocation,
+      highlightUpdatedAt,
       id,
-      source,
-      title,
       readwiseUrl,
+      source,
+      supertag,
+      title,
       url,
     } = values
     let t = '%%tana%%'
@@ -78,6 +81,15 @@ export default function Command() {
     highlights += highlightNote
       ? `{{#if note}}\n    - ${highlightNote}:: {{note}}{{/if}}`
       : '{{#if note}}\n    - **Note:** {{note}}{{/if}}'
+    highlights += highlightUpdatedAt
+      ? `{{#if updated}}\n    - ${highlightUpdatedAt}:: [[{{updated}}]]{{/if}}`
+      : ''
+    highlights += highlightHighlightedAt
+      ? `{{#if highlighted_at}}\n    - ${highlightHighlightedAt}:: [[{{highlighted_at}}]]{{/if}}`
+      : ''
+    highlights += highlightColor
+      ? `{{#if color}}\n    - ${highlightColor}:: {{color}}{{/if}}`
+      : ''
 
     highlights += '\n{{/each}}'